chore: Split non-chunk storage out of the ChunkStore
